♻️ refactor: moving init functions of ncurses to a separate file
